requirements.txt
flet==0.10.3
PyYAML~=6.0.1